+++ changelog.md
@@ -1,4 +1,8 @@
 ## 0.4.10
+- Upgrade docs/templates to use jquery-1.11.0.
+- Title is a link to the docs/index.html in service documentation pages.
+- Fix the missing jquery-min.map bug.
+- Fix validationSchema bugs in test templates.
 - Cleansing in the templates/project/server/api.js
 - Extend the server documentation:
     - Add references